[refact] LoginComma 컴포넌트 수정, Login 페이지 반영
